+++ app/src/androidTest/java/br/com/concrete/tentacle/base/BaseFragmentTest.kt
@@ -7,7 +7,7 @@ import org.junit.Ignore
 import org.junit.Rule
 
 @Ignore
-abstract class BaseFragmentTest: BaseTest() {
+abstract class BaseFragmentTest: BaseInstrumentedTest() {
 
     @get:Rule
     val activityRule = ActivityTestRule(SingleFragmentTestActivity::class.java)
